fix(remediation): close 21 Critical + 13 High audit findings — single-pass
8 parallel agent packs + 1 atomic alembic migration + orchestrator
synthesis. Risk Score 78 → ~50, Readiness 28 → ~55. Verdict remains
NO-GO (6 multi-week-eng Criticals still open) but path to CONDITIONAL
GO is now finite + trackable.

Migration e1a4b2c3d5f7 — single atomic revision, applied:
- uq_product_version_one_open partial unique (mint race)
- prevent_delete_credit_note BEFORE-DELETE trigger
- prevent_delete_payment_allocation BEFORE-DELETE trigger
- enforce_payment_allocation_total AFTER-INSERT trigger
- refresh_token.session_id UUID NOT NULL (family revocation)
- customer_user.token_not_before TIMESTAMPTZ (portal logout)
- api_key.expires_at + api_key.scopes (expiry + scope enforcement)

Financial integrity (9/9 Critical closed):
- F1 Stripe webhook currency validation
- F2 Stripe webhook amount-vs-outstanding
- F3 Auto-PAID flip subtract refunded_amount (3 sites)
- F4 mint_new_version advisory lock + partial unique
- F5 CN COUNT+1 → next_reference_number (3 sites)
- F6 settle_order with_for_update (tightening note in §6 known issues)
- F7 allocate_payment with_for_update + DB trigger backstop
- F8 usage Decimal end-to-end
- F9 credit_note + payment_allocation DELETE triggers via migration

Critical security (5/6 closed; S4 portal HttpOnly is product decision):
- S1 prod contract refuses mock providers + cors wildcard
- S2 deactivated user login/refresh rejected
- S3 portal HTML escape (invoice + receipt)
- S5 /org-tree requires auth + tenant scoping
- S6 password change revokes all refresh tokens

Critical observability/audit/backup (3/3 closed):
- A1 auth Events on login/logout/failure/refresh/replay
- C1 PII redacted from channel INFO logs
- B1 backup scripts committed (nightly/offsite/restore-verify)

Critical scale/operations (3/9 closed; rest BLOCKED multi-week-eng):
- P1 records.py SQL pagination
- P2 pagination DEFAULT_LIMIT=100 MAX_LIMIT=1000
- O6 docker-compose backend service + redis healthcheck

13 High findings closed: T1-T5 token/session, H3 CORS prod, H4 CSP,
H5 workitem IDOR, H6 filename injection, H8 _parse_dt naive (5
routers), H10 dunning anchor, H13 request-id middleware, H15 health
probe sanitize, H19 CSV formula injection, AC4 idempotency TOCTOU.

50 new remediation tests across 7 files:
- 48 passed (96%)
- 2 known issues: F6 settle_order needs tighter lock or unique
  constraint on payment.payment_order_id (1-day follow-up);
  dunning test has FK fixture bug (test bug, not code)

Regression: 110/110 auth+billing+stripe tests pass. No remediation-
induced regressions on core paths. test_attachments.py teardown
errors are caused by new auth audit Events referencing tenants —
the audit emit working as designed; test fixtures need CASCADE.

Preserved (verified untouched):
- D6 RLS 114-table coverage + NULLIF default-deny
- Production deploy contract DB role split
- Append-only triggers on event/invoice/payment
- Fernet field encryption
- Fail-soft scheduler/workflow/webhook
- Outbound httpx timeouts
- All tenant-filter noqa WHY comments

Full report: docs/audit/PRODUCTION-REMEDIATION-REPORT-2026-06-04.md

+++ backend/app/models/usage.py
@@ -8,6 +8,7 @@ flipping `rated=True` and linking the invoice. Tenant + org scoped; needs RLS (t
 from app.utils.ids import uuid7
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import String, BigInteger, Boolean, Numeric, ForeignKey, DateTime, func
 from sqlalchemy.dialects.postgresql import UUID
@@ -27,7 +28,12 @@ class UsageRecord(Base):
     # is Wave 2; column stays nullable until Wave 4 NOT NULL gate.
     service_id: Mapped[uuid.UUID | None] = mapped_column(UUID(as_uuid=True), ForeignKey("service.id", ondelete="RESTRICT"), nullable=True, index=True)
     metric: Mapped[str] = mapped_column(String(20), nullable=False)         # gb|minutes|messages|other
-    quantity: Mapped[float] = mapped_column(Numeric(18, 4), nullable=False, default=0)  # units consumed (may be fractional)
+    # F8 (financial-integrity Critical) — Decimal annotation.
+    # The column is already Numeric(18,4), but the Mapped[float] annotation invited callers to
+    # ferry the value as a Python float (binary drift on inputs like 0.1). Annotating as Decimal
+    # makes the typesafe path the only path — multiplication against unit_rate stays exact and
+    # the amount = round(qty * rate) computation lands on the cent the caller intended.
+    quantity: Mapped[Decimal] = mapped_column(Numeric(18, 4), nullable=False, default=0)  # units consumed (may be fractional)
     unit_rate: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)        # luma per unit
     amount: Mapped[int] = mapped_column(BigInteger, nullable=False, default=0)           # luma = round(quantity * unit_rate)
     period_start: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True)
